orbax/export/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "orbax-export"
version = "0.0.1"
description = "Orbax Export"
readme = "README.md"
license = "Apache-2.0"
authors = ["Orbax Authors <orbax-dev@google.com>"]
classifiers=[
    'Development Status :: 4 - Beta',
    'Intended Audience :: Developers',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: Apache Software License',
    'Programming Language :: Python',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
]
keywords = ["JAX machine learning", "serialization", "export"]
homepage = "http://github.com/google/orbax"
repository = "http://github.com/google/orbax"
packages = [
    { include = "../export" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
tensorflow = ">=2.9.3"
absl-py = "*"
etils = "*"
flax = "*"
jax = "*"
jaxlib = "*"
numpy = "*"
dataclasses-json = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-xdist = "*"